remove obsoleted code
