--- src/NServiceBus.Hosting.Azure.Tests/Hosting/When_checking_if_roleenvironment_is_available.cs
@@ -1,15 +1,0 @@
-﻿namespace NServiceBus.Hosting.Azure.Tests
-{
-    using NUnit.Framework;
-
-    [TestFixture]
-    [Category("Azure")]
-    public class When_checking_if_roleenvironment_is_available
-    {
-        [Test]
-        public void Should_not_throw_when_not_available()
-        {
-            Assert.AreEqual(false, SafeRoleEnvironment.IsAvailable);
-        }
-    }
-}
