--- src/NServiceBus.Hosting.Azure/ConfigurationSource/IAzureConfigurationSettings.cs
@@ -1,8 +1,0 @@
-namespace NServiceBus.Hosting.Azure
-{
-    interface IAzureConfigurationSettings
-    {
-        string GetSetting(string name);
-        bool TryGetSetting(string name, out string setting);
-    }
-}
