--- src/NServiceBus.Hosting.Azure/EndpointConfigurationExtensions.cs
@@ -1,20 +1,0 @@
-﻿namespace NServiceBus
-{
-    using Configuration.AdvancedExtensibility;
-
-    /// <summary>
-    /// Extensions for <see cref="EndpointConfiguration"/>.
-    /// </summary>
-    public static class EndpointConfigurationExtensions
-    {
-        /// <summary>
-        /// Defines the endpoint name in code thus overriding the previously assigned endpoint name.
-        /// </summary>
-        /// <param name="configuration">The configuration.</param>
-        /// <param name="endpointName">The endpoint name to be used.</param>
-        public static void DefineEndpointName(this EndpointConfiguration configuration, string endpointName)
-        {
-            configuration.GetSettings().Set("NServiceBus.Routing.EndpointName", endpointName);
-        }
-    }
-}
