--- src/NServiceBus.Hosting.Azure/IHost.cs
@@ -1,11 +1,0 @@
-namespace NServiceBus.Hosting.Azure
-{
-    interface IHost
-    {
-        void Start();
-
-        void Stop();
-
-        void Install(string username) ;
-    }
-}
