--- src/NServiceBus.Hosting.Azure/IsHostedIn.cs
@@ -1,15 +1,0 @@
-﻿namespace NServiceBus.Hosting.Azure
-{
-    using System.Diagnostics;
-
-    static class IsHostedIn
-    {
-        static string HostProcessName = "NServiceBus.Hosting.Azure.HostProcess";
-
-        public static bool ChildHostProcess()
-        {
-            var currentProcess = Process.GetCurrentProcess();
-            return currentProcess.ProcessName == HostProcessName;
-        }
-    }
-}
